feat: update `investmentlaunches migration` and `schema`

+++ db/migrate/20250904211140_create_investment_launches.rb
@@ -3,10 +3,10 @@
 class CreateInvestmentLaunches < ActiveRecord::Migration[7.1]
   def change
     create_table :investment_launches do |t|
-      t.integer :asset_type, null: false
-      t.integer :coin_type, null: false, default: 0
+      t.decimal :asset_type, null: false
+      t.decimal :coin_type, null: false, default: 0
       t.string :asset_name, null: false
-      t.integer :order_type, null: false
+      t.decimal :order_type, null: false
       t.date :order_date, null: false
       t.float :quantity, null: false
       t.float :unitary_price, null: false
